TSIS/TSIS1/phonebook_db/requirements.txt
psycopg2-binary==2.9.11
pygame==2.6.1
